processor/traffic-flow/requirements.txt
pulsar-client==2.4.0
opencv-python
argparse
numpy
mxnet
influxdb
configparser